refactor: 선수(투수,타자) 검색 기능을 hook으로 분리(#146)
