+++ src/features/player/OutfielderTab.tsx
@@ -1,20 +1,12 @@
-import { useSearchParams } from 'react-router';
 import Breadcrumb from '../common/Breadcrumb';
 import SearchBar from '../media/common/SearchBar';
 import { PlayerList } from './components';
 import NotFonudSearch from './components/NotFonudSearch';
-import { usePlayerList } from './hooks/usePlayerList';
+import { usePlayerSearch } from './hooks/usePlayerSearch';
 
 function OutfielderTab() {
-  const { playerList, loading, error } = usePlayerList('outfielder');
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  const searchWord = searchParams.get('searchWord') || '';
-
-  // 검색 결과 필터링
-  const filteredOutfielderList = playerList.filter((player) =>
-    player.playerName.toLowerCase().includes(searchWord.toLowerCase())
-  );
+  const { filteredPlayerList, loading, error, searchWord, handleSearch } =
+    usePlayerSearch('outfielder');
 
   if (error) {
     return <div>Error: {error}</div>;
@@ -23,23 +15,13 @@ function OutfielderTab() {
   return (
     <>
       <Breadcrumb
-        leftComponent={
-          <SearchBar
-            value={searchParams.get('searchWord') || ''}
-            onSubmit={(searchWord) =>
-              setSearchParams({
-                ...Object.fromEntries(searchParams.entries()),
-                searchWord,
-              })
-            }
-          />
-        }
+        leftComponent={<SearchBar value={searchWord} onSubmit={handleSearch} />}
       />
-      {!loading && !error && filteredOutfielderList.length === 0 ? (
+      {!loading && !error && filteredPlayerList.length === 0 ? (
         <NotFonudSearch />
       ) : (
         <PlayerList
-          playerList={error ? [] : filteredOutfielderList}
+          playerList={error ? [] : filteredPlayerList}
           endpoint="outfielder"
           loading={loading}
         />
