+++ src/pages/player/PitcherPage.tsx
@@ -3,19 +3,11 @@ import Breadcrumb from '@/features/common/Breadcrumb';
 import SearchBar from '@/features/media/common/SearchBar';
 import { PlayerList } from '@/features/player/components';
 import NotFonudSearch from '@/features/player/components/NotFonudSearch';
-import { usePlayerList } from '@/features/player/hooks/usePlayerList';
-import { useSearchParams } from 'react-router';
+import { usePlayerSearch } from '@/features/player/hooks/usePlayerSearch';
 
 function PitcherPage() {
-  const { playerList, loading, error } = usePlayerList('pitcher');
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  const searchWord = searchParams.get('searchWord') || '';
-
-  // 검색 결과 필터링
-  const filteredPlayerList = playerList.filter((coach) =>
-    coach.playerName.toLowerCase().includes(searchWord.toLowerCase())
-  );
+  const { filteredPlayerList, loading, error, searchWord, handleSearch } =
+    usePlayerSearch('pitcher');
 
   if (error) {
     return <div>Error: {error}</div>;
@@ -34,17 +26,7 @@ function PitcherPage() {
         </Banner.Overlay>
       </Banner>
       <Breadcrumb
-        leftComponent={
-          <SearchBar
-            value={searchParams.get('searchWord') || ''}
-            onSubmit={(searchWord) =>
-              setSearchParams({
-                ...Object.fromEntries(searchParams.entries()),
-                searchWord,
-              })
-            }
-          />
-        }
+        leftComponent={<SearchBar value={searchWord} onSubmit={handleSearch} />}
       />
       {!loading && !error && filteredPlayerList.length === 0 ? (
         <NotFonudSearch />
